requirements.txt
yfinance
kafka-python
psycopg2
dotenv
streamlit 
psycopg2-binary 
pandas